Add register flood protection (closed #263)
Adds the rcon command to manually ban ips from using the register command.
And also allows to set a maximum amount of accounts that are allowed per ip before banning the ip from account creation.

If more than 128 people are tracked in the system the server starts increase the human level needed to register the account.
So people have to play active wait x seconds do quests and get blockpoints to create an account.