Remove SvHaxx0rSpoof due to big security risk
I know removing features that don't crash the server is against the bloated ddnet++ philosophy.
But version 007 is planned as the most stable and secure version ever released of ddnet++.

Even if it is a admin only risk. It allows for compromization of the whole machine the server is running on.
Not even I trust my admins enough to not fuck this up.. and I have no admins besides me xd.

So to protect users of this modification this feature is removed for now.