Improve hotfix for broken accounts (closed #246)
Somehow accounts to get stuck when they are marked as logged in in the database but actually arent.
To solve this sql_logout_all was implemented. But it possible marked players as logged out in the database while being logged in.

So this commit solves the issue by only logging out players that logged in on the current port and are not currently online.
So if the port does not change this is pretty save. That is why there is also an cfg to automatically do this every few ticks.
With the config activated nobody should need a supporter to be able to login again. Still a hack that only hotfixes the result of the bug
and not the core of the bug it self but it works.

sql_logout <name> is still very dangerous tho. I don't know if it is usefull probably should be removed.