Remove bugged m_aClientLeftTheServer (Closed #367)
It was never set to false again.
So as soon as someone leaves that ID would be marked as left.
Checking m_apPlayers works with less code.

But there still might be a race condition with fast reconnects.
The weapon drop cleanup should happen in OnPlayerDisconnect instead of CWeapon::Tick()

But lets not change all code for now and see if it works.