++ remove PreTickDDPP
The regular character ddpp tick is also called from pre tick.
So this was a PrePreTick. But it was not a reliable PrePreTick.
Because ddnets PreTick has two call sites depending on weak hook
being on or off.

Meaning the moneytile reset would be initialized wrongly if weak hook is off.
And in general its just confusing. And not needed.

There is no reason this had to be done outside of the regular tick.
The only reason could be code readability but then the PreTickDDPP
could be called from the DDPP_Tick instead of from the ddnet code.
But for now I do not think that is needed.