++ remove old unused score fix for ddnet
Not sure what it did exactly. I think @fokkonaut built that back in the
days. Probably to better represent non time values in the ddrace
scoreboard.

So when we want to display the level 6 ddnet clients would see the time
value 00:06 or something like that.

I assume this entire feature broke in a merge. Because the variable is
never set. Nowerdays the use of old ddnet clients dropped to irrelevant
values. Players either use pure teeworlds or ddnet that is new enough
to properly support score flags where we can display an actual 6 instead
of hacking in a 00:06. This makes the code simpler and drops support for
these old clients which is fair imo.