++ naive implementation of flag drop tile
Closed #430
Relevant for #458 because no hammer and captcha tiles got shifted one to
the right to make space for the flag drop tile.

So we can use the F-DDrace compatible index 185

The implementation is naive because even if the tile name says STOP
one is actually not stopped from moving a flag through the tile.
Because the flag can be dropped through the tile.