++ track which server ip the accounts are logged in
This is crucial for multi host support to not corrupt the database.
If we want to cleanup accounts that got stuck in logged in state
we can not use the port as unique identifier because multiple
hosts could share the same port.

Also when trying to alter accounts via an admin command it will
be useful to know which server the account is logged in.