++ disallow editing accounts logged in on other servers
Fixes state being reverted when an admin tries to edit
an account that is currently playing on another host or port.
In that case the logout of the account could overwrite the state
again. So we show a useful error to the admin instead of claiming
something got applied that will later be reverted.